sprint/5-step-1 - Привел в порядок расположение файлов - Переместил action creator's в папку action - Добавил типизированые хуки - Заменил во всем проекте не типизированные хуки на типизированные - Типизировал состояние редакса
