+++ src/compontents/hoc/ProtectedAuthRoute.tsx
@@ -1,7 +1,6 @@
 import React, {FC, ReactElement} from 'react';
 import {Navigate, useLocation} from "react-router-dom";
-import {useSelector} from "react-redux";
-import {RootState} from "../../services/reducers";
+import {useSelector} from "../../services/hooks";
 
 interface IProtectedUnAuthRouteProps {
     children: ReactElement
@@ -9,10 +8,8 @@ interface IProtectedUnAuthRouteProps {
 
 const ProtectedUnAuthRoute:FC<IProtectedUnAuthRouteProps> = ({children}) => {
     const location = useLocation()
-    const {isAuth} = useSelector((state:RootState) => state.auth)
+    const {isAuth} = useSelector(state => state.auth)
     return isAuth ? <Navigate to='/' state={{from: location}}/> : children
-
-
 };
 
 export default ProtectedUnAuthRoute;
