+++ src/compontents/hoc/ProtectedUnAuthRoute.tsx
@@ -1,7 +1,6 @@
 import React, {ReactElement, FC} from 'react';
 import {Navigate, useLocation} from "react-router-dom";
-import {useSelector} from "react-redux";
-import {RootState} from "../../services/reducers";
+import {useSelector} from "../../services/hooks";
 
 interface IProtectedUnAuthRoute {
     children: ReactElement
@@ -9,7 +8,7 @@ interface IProtectedUnAuthRoute {
 
 const ProtectedUnAuthRoute:FC<IProtectedUnAuthRoute> = ({children}) => {
     const location = useLocation()
-    const {isAuth} = useSelector((state:RootState) => state.auth)
+    const {isAuth} = useSelector(state => state.auth)
     return !isAuth ? <Navigate to='/login' state={{from: location}}/> : children
 };
