+++ src/compontents/modals/Modal/Modal.tsx
@@ -4,9 +4,10 @@ import ModalStyles from "./Modal.module.scss";
 import {CloseIcon} from "@ya.praktikum/react-developer-burger-ui-components";
 import ModalOverlay from "../ModalOverlay/ModalOverlay";
 import {CSSTransition} from "react-transition-group";
-import {useDispatch, useSelector} from "react-redux";
-import {RootState} from "../../../services/reducers";
-import { deleteAllIngredientsFromCart } from '../../../services/reducers/burgerCounstructor';
+
+import { deleteAllIngredientsFromCart } from '../../../services/actions/burgerCounstructor';
+import {useDispatch, useSelector} from "../../../services/hooks";
+
 
 interface IModal {
     onCloseModal: () => void
@@ -16,7 +17,7 @@ const Modal: FC<IModal> = ({children, onCloseModal}) => {
     const [containerState, setContainerState] = useState(false)
     const portalDiv = document.getElementById('modals')!
     const nodeRef = React.useRef(null)
-    const orderState = useSelector((state:RootState) => state.order)
+    const orderState = useSelector(state => state.order)
     const dispatch = useDispatch()
 
     const closeModal = () => {
