+++ src/services/actions/auth.ts
@@ -1,4 +1,4 @@
-import {deleteCookie, setCookie} from "../../utils/helpers";
+import {deleteCookie, setCookie} from "../helpers";
 
 import {
     fetchRegisterRequest,
@@ -8,24 +8,24 @@ import {
     fetchLogOut,
     fetchForgotPasswordRequest,
     fetchUpdateUser
-} from "../../utils/api";
+} from "../api";
 
 import {
     IForgotPasswordUserTypes,
     ILoginUserTypes,
     IRegisterUserTypes, IResetPasswordTypes,
-    IUpdateUserTypes
+    IUpdateUserTypes, IUserTypes
 } from "../../types/ingredientTypes";
 
-import {
-    deleteAuth, loadingUserComplete,
-    loginFailed,
-    loginOrPasswordIncorrect,
-    loginSuccess, registerUserClearTextError,
-    registerUserFailed, registerUserSetTextError,
-    registerUserSuccess, resetPasswordComplete, resetPasswordSuccess, setUserInfo
-} from "../reducers/auth";
 import {AppDispatch, AppThunk} from "../../types";
+import {
+    IDeleteAuth, ILoadingUser, ILoadingUserComplete,
+    ILoginFailed,
+    ILoginOrPasswordIncorrect,
+    ILoginSuccess, IRegisterUserClearTextError,
+    IRegisterUserFailed, IRegisterUserSetTextError,
+    IRegisterUserSuccess, IResetPasswordComplete, IResetPasswordSuccess, ISetUserInfo
+} from "../../types/authTypes";
 
 
 export const REGISTER_USER_FAILED: 'REGISTER_USER_FAILED' = 'REGISTER_USER_FAILED'
@@ -95,7 +95,7 @@ export const registerRequest: AppThunk = (form: IRegisterUserTypes, callback: ()
         })
 };
 
-export const forgotPassword = (form: IForgotPasswordUserTypes) => (dispatch: any): void => {
+export const forgotPassword: AppThunk = (form: IForgotPasswordUserTypes) => (dispatch: AppDispatch) => {
     fetchForgotPasswordRequest(form)
         .then(response => {
             if (!response.ok) {
@@ -117,7 +117,7 @@ export const forgotPassword = (form: IForgotPasswordUserTypes) => (dispatch: any
         })
 }
 
-export const resetPassword = (form: IResetPasswordTypes) => (dispatch: any): void => {
+export const resetPassword: AppThunk = (form: IResetPasswordTypes) => (dispatch: AppDispatch) => {
     fetchResetPasswordRequest(form)
         .then(response => {
             if (response.message === 'Password successfully reset') {
@@ -129,7 +129,7 @@ export const resetPassword = (form: IResetPasswordTypes) => (dispatch: any): voi
         })
 }
 
-export const getUserInfo = () => (dispatch: any) => {
+export const getUserInfo: AppThunk = () => (dispatch: AppDispatch) => {
     //TODO Сделать проверку на отсутствие refreshToken
     //TODO При его отсутсвтии ломается приложение
 
@@ -147,7 +147,7 @@ export const getUserInfo = () => (dispatch: any) => {
         })
 }
 
-export const logOutUser = (callback: () => void) => (dispatch: any) => {
+export const logOutUser: AppThunk = (callback: () => void) => (dispatch: AppDispatch) => {
     fetchLogOut()
         .then(response => {
             if (response.ok) {
@@ -171,7 +171,7 @@ export const logOutUser = (callback: () => void) => (dispatch: any) => {
         })
 }
 
-export const updateUserInfo = (form: IUpdateUserTypes, password?:string) => () => {
+export const updateUserInfo: AppThunk = (form: IUpdateUserTypes, password?:string) => () => {
     fetchUpdateUser(form, password)
         .then(response => {
             console.log(response)
@@ -180,3 +180,60 @@ export const updateUserInfo = (form: IUpdateUserTypes, password?:string) => () =
             throw new Error(e)
         })
 }
+
+export const loginSuccess = (user: IUserTypes): ILoginSuccess => ({
+    type: LOGIN_SUCCESS,
+    user: user
+})
+
+export const loginFailed = (): ILoginFailed=> ({
+    type: LOGIN_FAILED
+});
+
+export const loginOrPasswordIncorrect = (message: string): ILoginOrPasswordIncorrect => ({
+    type: LOGIN_OR_PASSWORD_INCORRECT,
+    message: message
+})
+
+export const registerUserFailed = (): IRegisterUserFailed => ({
+    type: REGISTER_USER_FAILED
+})
+
+export const registerUserSuccess = (user: IUserTypes): IRegisterUserSuccess => ({
+    type: REGISTER_USER_SUCCESS,
+    payload: user
+})
+
+export const registerUserSetTextError = (message: string): IRegisterUserSetTextError => ({
+    type: REGISTER_USER_SET_TEXT_ERROR,
+    message: message,
+})
+
+export const registerUserClearTextError = (): IRegisterUserClearTextError => ({
+    type: REGISTER_USER_CLEAR_TEXT_ERROR
+})
+
+export const resetPasswordSuccess = (): IResetPasswordSuccess => ({
+    type: RESET_PASSWORD_SUCCESS
+})
+
+export const resetPasswordComplete = (): IResetPasswordComplete=> ({
+    type: RESET_PASSWORD_COMPLETED
+})
+
+export const setUserInfo = (userInfo: IUserTypes): ISetUserInfo => ({
+    type: SET_USER_INFO,
+    payload: userInfo
+})
+
+export const deleteAuth = (): IDeleteAuth => ({
+    type: DELETE_AUTH
+})
+
+export const loadingUser = (): ILoadingUser => ({
+    type: LOADING_USER
+})
+
+export const loadingUserComplete = (): ILoadingUserComplete => ({
+    type: LOADING_USER_COMPLETED
+})
