+++ src/services/actions/burgerIngredients.ts
@@ -1,15 +1,18 @@
-import {API_REACT} from "../../utils/data";
+import {API_REACT} from "../url";
+
+import {AppDispatch, AppThunk} from "../../types";
 import {
-    getAllIngredientsFailed,
-    getAllIngredientsRequest,
-    getAllIngredientsSuccess
-} from "../reducers/burgerIngredients";
+    IGetAllIngredientsFailed,
+    IGetAllIngredientsRequest,
+    IGetAllIngredientsSuccess
+} from "../../types/burgerIngredientsTypes";
+import {IIngredient} from "../../types/ingredientTypes";
 
 export const GET_ALL_INGREDIENTS_SUCCESS: 'GET_ALL_INGREDIENTS_SUCCESS' = 'GET_ALL_INGREDIENTS_SUCCESS'
 export const GET_ALL_INGREDIENTS_REQUEST: 'GET_ALL_INGREDIENTS_REQUEST' = 'GET_ALL_INGREDIENTS_REQUEST'
 export const GET_ALL_INGREDIENTS_FAILED: 'GET_ALL_INGREDIENTS_FAILED' = 'GET_ALL_INGREDIENTS_FAILED'
 
-export const fetchIngredients = () => (dispatch: any) => {
+export const fetchIngredients: AppThunk = () => (dispatch: AppDispatch) => {
     dispatch(getAllIngredientsRequest())
 
     fetch(API_REACT + '/ingredients')
@@ -29,3 +32,15 @@ export const fetchIngredients = () => (dispatch: any) => {
         })
 }
 
+export const getAllIngredientsRequest = (): IGetAllIngredientsRequest => ({
+    type: GET_ALL_INGREDIENTS_REQUEST
+})
+
+export const getAllIngredientsSuccess = (ingredients: IIngredient[]): IGetAllIngredientsSuccess => ({
+    type: GET_ALL_INGREDIENTS_SUCCESS,
+    ingredients: ingredients
+})
+
+export const getAllIngredientsFailed = (): IGetAllIngredientsFailed => ({
+    type: GET_ALL_INGREDIENTS_FAILED
+})
