+++ src/services/reducers/burgerIngredients.ts
@@ -5,16 +5,7 @@ import {
 
 } from "../actions/burgerIngredients";
 
-import {IIngredient} from "../../types/ingredientTypes";
-
-import {
-    IGetAllIngredientsFailed,
-    IGetAllIngredientsRequest, IGetAllIngredientsSuccess,
-    IIngredientsState,
-    TBurgerIngredientsActions
-} from "../../types/burgerIngredientsTypes";
-
-
+import {IIngredientsState, TBurgerIngredientsActions} from "../../types/burgerIngredientsTypes";
 
 const defaultState: IIngredientsState = {
     ingredients: [],
@@ -43,15 +34,3 @@ export const ingredientsReducer = (state = defaultState, action: TBurgerIngredie
     }
 }
 
-export const getAllIngredientsRequest = (): IGetAllIngredientsRequest => ({
-    type: GET_ALL_INGREDIENTS_REQUEST
-})
-
-export const getAllIngredientsSuccess = (ingredients: IIngredient[]): IGetAllIngredientsSuccess => ({
-    type: GET_ALL_INGREDIENTS_SUCCESS,
-    ingredients: ingredients
-})
-
-export const getAllIngredientsFailed = (): IGetAllIngredientsFailed => ({
-    type: GET_ALL_INGREDIENTS_FAILED
-})
